added traverse part

+++ tests/srcs/ArithmeticGrammar.cpp
@@ -6,7 +6,7 @@
 /*   By: ldedier <ldedier@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/28 09:57:18 by ldedier           #+#    #+#             */
-/*   Updated: 2020/01/04 16:00:12 by ldedier          ###   ########.fr       */
+/*   Updated: 2020/01/04 23:57:42 by ldedier          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ std::deque<Token<int, int> *> ArithmeticGrammar::innerLex(std::istream & istream
 		else if (c == '-')
 			res.push_back(new Token<int, int>(*(getTerminal("-"))));
 		else if (isdigit(c))
-			res.push_back(new Token<int, int>(*(getTerminal("integer")), c - '0'));
+			res.push_back(new Token<int, int>(*(getTerminal("integer")), (int)(c - '0')));
 		else if (!isspace(c))
 			throw std::exception();
 	}
